Update List.java
Hi Justin, I'm a big fan of your code, but, I think I may have spotted an issue in your add method for ArrayList. When the index of the element that you want to insert is smaller than the size of the list, you shift the array down one spot and insert the value at the desired position. However, I think you forgot to increase the size of the list after that, which may cause some problems when iterating through the arrayList.