Master Comments BubbleSort.java
<!--
Hi!
Thanks for considering contributing to this ever-growing list of algorithm and data structure implementations in java.
Your contribution is valuable.
In order to help us evaluate PRs better, we ask you to have a look at the following declaration and check the points you agree with. ( [x] )
PRs which don't agree to all the points mentioned below will be rejected.
-->

#### By submitting this pull request I confirm I've read and complied with the below requirements.

- [ ] I have read the [Contribution guidelines](CONTRIBUTING.md) and I am confident that my PR reflects them.
- [ ] I have followed the [coding guidelines](CONTRIBUTING.md#cs) for this project.
- [ ] My code follows the [skeleton code structure](CONTRIBUTING.md#sample).
- [ ] This pull request has a descriptive title. For example, `Added {Algorithm/DS name} [{Language}]`, not `Update README.md` or `Added new code`.